Move Under File System tests to tachyon.underfs package (TACHYON-221)
Relocates test code for Under File System into the relevant package
