+++ core/src/test/java/tachyon/underfs/UnderFileSystemCluster.java
@@ -12,12 +12,14 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package tachyon;
+package tachyon.underfs;
 
 import java.io.IOException;
 
 import com.google.common.base.Throwables;
 
+import tachyon.LocalFilesystemCluster;
+import tachyon.TachyonURI;
 import tachyon.underfs.UnderFileSystem;
 import tachyon.util.CommonUtils;
