+++ core/src/test/java/tachyon/underfs/UnderFileSystemTest.java
@@ -12,11 +12,13 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package tachyon;
+package tachyon.underfs;
 
 import org.junit.Assert;
 import org.junit.Test;
 
+import tachyon.Pair;
+import tachyon.TachyonURI;
 import tachyon.underfs.UnderFileSystem;
 
 /**
