+++ core/src/test/java/tachyon/underfs/UnderFileSystemsUtils.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package tachyon;
+package tachyon.underfs;
 
 import java.io.IOException;
